# src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "REST API minta backend",
        "version": "1.0.0",
        "description": "<img alt='DB_diagram' height='700px' src='https://nitslaszlo.github.io/static/session/db_diagram_new.jpg' />"
    },
    "servers": [
        {
            "url": "https://jedliksession.cyclic.app"
        },
        {
            "url": "http://localhost:5000"
        }
    ],
    "paths": {
...
        "/recipes": {
            "get": {
                "tags": [
                    "recipes"
                ],
                "summary": "Összes recept lekérdezése.",
                "description": "Az összes recept adatát tudjuk lekérdezni.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/recipe"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorSchema"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "recipes"
                ],
                "summary": "Új recept mentése.",
                "description": "Az új recept mentése a 'hirdetesek' táblában",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "$ref": "#/components/schemas/recipe"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK (Az új, elmentett recepttel tér vissza.)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/recipe"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/recipes/{offset}/{limit}/{field}": {
            "get": {
                "tags": [
                    "recipes"
                ],
                "summary": "Paginált adatok lekérése.",
                "description": "receptek lekérdezése lapozáshoz.",
                "parameters": [
                    {
                        "name": "offset",
                        "description": "Ennyi rekordot hagyunk el (skip) a lekérdezés 'elejéről'.",
                        "in": "path",
                        "schema": {
                            "type": "number",
                            "example": "5"
                        },
                        "required": true
                    },
                    {
                        "name": "limit",
                        "description": "Ennyi rekordot küld (take) maximum az API",
                        "in": "path",
                        "schema": {
                            "type": "number",
                            "example": "5"
                        },
                        "required": true
                    },
                    {
                        "name": "field",
                        "description": "Melyik mező szerint legyenek rendezve a rekordok. A kötőjel '-' karakter a csökkenő rendezést jelöli.",
                        "in": "path",
                        "schema": {
                            "type": "string",
                            "example": "-cim"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "count": {
                                            "type": "number",
                                            "description": "Az összes rekord száma.",
                                            "example": "16"
                                        },
                                        "data": {
                                            "type": "array",
                                            "description": "A maximum {limit} számú rekord.",
                                            "items": {
                                                "$ref": "#/components/schemas/recipe"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorSchema"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/recipes/{offset}/{limit}/{field}/{filter}": {
            "get": {
                "tags": [
                    "recipes"
                ],
                "summary": "Paginált adatok lekérése szűréssel.",
                "description": "Receptek lekérdezése lapozáshoz szűréssel.",
                "parameters": [
                    {
                        "name": "offset",
                        "description": "Ennyi rekordot hagyunk el (skip) a lekérdezés 'elejéről'.",
                        "in": "path",
                        "schema": {
                            "type": "number",
                            "example": "0"
                        },
                        "required": true
                    },
                    {
                        "name": "limit",
                        "description": "Ennyi rekordot küld (take) maximum az API.",
                        "in": "path",
                        "schema": {
                            "type": "number",
                            "example": 2
                        },
                        "required": true
                    },
                    {
                        "name": "field",
                        "description": "Melyik mező szerint legyenek rendezve a rekordok. A kötőjel '-' karakter a csökkenő rendezést jelöli.",
                        "in": "path",
                        "schema": {
                            "type": "string",
                            "example": "-cim"
                        },
                        "required": true
                    },
                    {
                        "name": "filter",
                        "description": "Szűrő - a 'cim' és a 'leiras' mezőkben keressük.",
                        "in": "path",
                        "schema": {
                            "type": "string",
                            "example": "leves"
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "count": {
                                            "type": "number",
                                            "description": "A szűrőnek megfelelő rekordok száma",
                                            "example": "16"
                                        },
                                        "data": {
                                            "type": "array",
                                            "description": "A maximum {limit} számú, szűrőnek megfelelő rekord.",
                                            "items": {
                                                "$ref": "#/components/schemas/recipe"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorSchema"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "security": [
        {
            "name": []
        }
    ],
    "components": {
        "schemas": {
...
            "recipe": {
                "type": "object",
                "required": [
                    "recipeName"
                ],
                "properties": {
                    "_id": {
                        "type": "string",
                        "description": "PK - A recept azonosítója (ObjectID).",
                        "example": "61daf48ba5775848e99f555b"
                    },
                    "user_id": {
                        "type": "string",
                        "description": "FK - A recept szerzője (ObjectID).",
                        "example": "61b5e9c0f39e4edcf5b8a3b9"
                    },
                    "recipeName": {
                        "type": "string",
                        "maxLength": 50,
                        "description": "A recept neve.",
                        "example": "PENNE ALL ARRABBIATA CSIRKEMELLEL"
                    },
                    "imageURL": {
                        "type": "string",
                        "description": "Az elkészült ételt ábrázoló kép URL címe.",
                        "example": "https://www.mindmegette.hu/images/240/O/penne-arrabbiata-csirkes.jpg"
                    },
                    "decription": {
                        "type": "string",
                        "description": "Az étel elkészítésének leírása.",
                        "example": "A tésztát a csomágolásán látható utasítás szerint forró, sós vízbe kifőzzük és leszűrjük. A finomra vágott hagymákat olajon megdinszteljük, majd hozzáadjuk a kockára vágott csirkemellet és fehéredésig pirítjuk. Sózzuk, borsozzuk, ráöntjük a paradicsomkonzervet és a vörösbort majd forrásig melegítjük.\nHozzáadjuk a chilit, a felaprított bazsalikomot, az oregánót, végül pedig a barnacukrot és 25-30 percig főzzük, míg a csirkemell meg nem puhul, majd összekeverjük a tésztával."
                    },
                    "ingredients": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "A hozzávalókat tartalmazó tömb.",
                        "example": "[\"50 dkg penne\", \"1 db vöröshagyma\", \"4 gerezd fokhagyma\", \"30 dkg\tcsirkemell\", \"só\", \"bors\", \"stb.\"]"
                    }
                }
            },
...
            "ErrorSchema": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "error.message value"
                    },
                    "status": {
                        "type": "integer",
                        "example": 400
                    }
                }
            }
        }
    }
}